expenzo/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
Faker==11.3.0
importlib-metadata==4.8.2
Markdown==3.3.6
mysqlclient==2.1.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uWSGI==2.0.20
zipp==3.6.0
